# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "订单微服务相关接口",
        "title": "订单服务 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/order-ms/v1",
    "paths": {
        "/customer/create": {
            "post": {
                "description": "创建一个新订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.OrderInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "types.OrderInfo": {
            "type": "object",
            "properties": {
                "order_item_list": {
                    "description": "订单商品列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.OrderItemInfo"
                    }
                },
                "receiver_address": {
                    "description": "收货地址",
                    "type": "string"
                },
                "receiver_country": {
                    "description": "收货人国家",
                    "type": "string"
                },
                "receiver_first_name": {
                    "description": "收货人姓名",
                    "type": "string"
                },
                "receiver_last_name": {
                    "description": "收货人姓名",
                    "type": "string"
                },
                "receiver_phone": {
                    "description": "收货人电话",
                    "type": "string"
                },
                "receiver_zip_code": {
                    "description": "收货人邮政编码",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                }
            }
        },
        "types.OrderItemInfo": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        }
    }
}